requirements.txt
Flask==1.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.14.0
SQLAlchemy==1.3.20
twilio==6.45.4
